requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
kavenegar>=1.1.2
Pillow>=10.0,<11.0
